+++ src/Query/Criteria/Filter.php
@@ -5,7 +5,7 @@ namespace Kununu\Elasticsearch\Query\Criteria;
 
 final class Filter extends AbstractFilter
 {
-    public static function create(string $field, mixed $value, ?string $operator = null, array $options = []): Filter
+    public static function create(string $field, mixed $value, ?string $operator = null, array $options = []): self
     {
         return new self($field, $value, $operator, $options);
     }